+++ NEWS.md
@@ -1,3 +1,5 @@
+# resourcecode (development version)
+
 # resourcecode 0.5.3
 
 - Custom labels are now handled properly in `cut_seasons()` ;