+++ NEWS.md
@@ -1,4 +1,4 @@
-# resourcecode (development version)
+# resourcecode 0.5.4
 
 - `get_parameters` now correctly handles `Date` format as inputs by spanning on the right period;
 - Update `rscd_casandra_end_date` with new data availability: 2025-12-31 23:00:00.